# openapi.json
{"openapi": "3.0.2", "info": {"title": "mssql-connector", "version": "0.1.0"}, "paths": {"/metrics": {"get": {"summary": "Metrics", "description": "Endpoint that serves Prometheus metrics.", "operationId": "metrics_metrics_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/specification/list": {"get": {"tags": ["connector specification endpoints"], "summary": "Get Connector Spec", "operationId": "get_connector_spec_api_v1_specification_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Connector Spec Api V1 Specification List Get"}}}}}}}, "/api/v1/specification/test": {"post": {"tags": ["connector specification endpoints"], "summary": "Test Connector Spec", "operationId": "test_connector_spec_api_v1_specification_test_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Test Connector Spec Api V1 Specification Test Post"}}}}}}}, "/api/v1/db_tables/list": {"get": {"tags": ["table metadata"], "summary": "Get Tables", "operationId": "get_tables_api_v1_db_tables_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Tables Api V1 Db Tables List Get", "type": "array", "items": {}}}}}}}}, "/api/v1/db_tables/get/{name}": {"get": {"tags": ["table metadata"], "summary": "Get Table Columns", "operationId": "get_table_columns_api_v1_db_tables_get__name__get", "parameters": [{"required": true, "schema": {"title": "Name", "type": "string"}, "name": "name", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Table Columns Api V1 Db Tables Get  Name  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/schema/list": {"get": {"tags": ["database schema"], "summary": "Get Schemas", "operationId": "get_schemas_api_v1_schema_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Schemas Api V1 Schema List Get", "type": "array", "items": {}}}}}}}}, "/api/v1/rpc/list": {"get": {"tags": ["stored procedures"], "summary": "Get User Defined Function", "description": "Create or replace user defined functions", "operationId": "get_user_defined_function_api_v1_rpc_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get User Defined Function Api V1 Rpc List Get"}}}}}}}, "/api/v1/rpc/create": {"post": {"tags": ["stored procedures"], "summary": "Create User Defined Function", "description": "Create or replace user defined functions", "operationId": "create_user_defined_function_api_v1_rpc_create_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/NewFunction"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Create User Defined Function Api V1 Rpc Create Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/rpc/{name}": {"post": {"tags": ["stored procedures"], "summary": "Execute Rpc Function", "description": "Execute user defined function\n:param request:\n:param name: name of the function\n:param session:\n:return: Tuple output", "operationId": "execute_rpc_function_api_v1_rpc__name__post", "parameters": [{"required": true, "schema": {"title": "Name", "type": "string"}, "name": "name", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Execute Rpc Function Api V1 Rpc  Name  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["stored procedures"], "summary": "Delete User Defined Function", "description": "Delete user defined functions", "operationId": "delete_user_defined_function_api_v1_rpc__name__delete", "parameters": [{"required": true, "schema": {"title": "Name", "type": "string"}, "name": "name", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Delete User Defined Function Api V1 Rpc  Name  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{table}": {"get": {"tags": ["table records"], "summary": "Read Records", "operationId": "read_records_api_v1__table__get", "parameters": [{"required": true, "schema": {"title": "Table", "type": "string"}, "name": "table", "in": "path"}, {"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 100}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Read Records Api V1  Table  Get", "type": "array", "items": {}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["table records"], "summary": "Create Record", "description": "Create new record.", "operationId": "create_record_api_v1__table__post", "parameters": [{"required": true, "schema": {"title": "Table", "type": "string"}, "name": "table", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Create Record Api V1  Table  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["table records"], "summary": "Delete Record", "description": "Delete record by supplying variable parameters (similar to where clause) i.e. key1=value1&key2=value2", "operationId": "delete_record_api_v1__table__delete", "parameters": [{"required": true, "schema": {"title": "Table", "type": "string"}, "name": "table", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Delete Record Api V1  Table  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{table}/query": {"get": {"tags": ["table records"], "summary": "Read Records By Query Params", "description": "Read records by supplying variable parameters (similar to where clause) i.e. key1=value1&key2=value2", "operationId": "read_records_by_query_params_api_v1__table__query_get", "parameters": [{"required": true, "schema": {"title": "Table", "type": "string"}, "name": "table", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Read Records By Query Params Api V1  Table  Query Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/{table}/bulk": {"post": {"tags": ["table records"], "summary": "Insert Bulk Records", "description": "Insert bulk records.", "operationId": "insert_bulk_records_api_v1__table__bulk_post", "parameters": [{"required": true, "schema": {"title": "Table", "type": "string"}, "name": "table", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Items"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Insert Bulk Records Api V1  Table  Bulk Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/query/raw/sql": {"post": {"tags": ["table records"], "summary": "Run Sql Command", "description": "Run raw sql command.", "operationId": "run_sql_command_api_v1_query_raw_sql_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RawSQL"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Run Sql Command Api V1 Query Raw Sql Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Item": {"title": "Item", "required": ["fields"], "type": "object", "properties": {"fields": {"$ref": "#/components/schemas/Row"}}}, "Items": {"title": "Items", "required": ["fields"], "type": "object", "properties": {"fields": {"title": "Fields", "type": "array", "items": {"$ref": "#/components/schemas/Row"}}}}, "NewFunction": {"title": "NewFunction", "required": ["function_body"], "type": "object", "properties": {"function_body": {"title": "Function Body", "type": "string"}}}, "RawSQL": {"title": "RawSQL", "type": "object", "properties": {"raw_sql": {"title": "Raw Sql", "type": "string"}}}, "Row": {"title": "Row", "type": "object", "properties": {}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}